models removed to report
