+++ src/sklearn_evaluation/report/__init__.py
@@ -1,6 +1,14 @@
 __version__ = "0.11.2dev"
 
 from sklearn_evaluation.report.model_heuristics import ModelHeuristics, ReportSection
+from sklearn_evaluation.report.model_evaluator import ModelEvaluator, evaluate_model
+from sklearn_evaluation.report.models_comparer import ModelsComparer, compare_models
 
 
-__all__ = ["ModelHeuristics", "ReportSection"]
+__all__ = ["ModelHeuristics",
+           "ReportSection",
+           "ModelEvaluator",
+           "evaluate_model",
+           "compare_models",
+           "ModelsComparer"
+           ]
