+++ src/sklearn_evaluation/report/util.py
@@ -1,4 +1,5 @@
 from jinja2 import Environment, PackageLoader
+import matplotlib.pyplot as plt
 
 
 def jinja_env():
@@ -35,3 +36,23 @@ def run_if_args_are_not_none(func):
             return func(*args, **kw)
 
     return wrapper
+
+
+def gen_ax():
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    return ax
+
+
+def check_model(model) -> None:
+    """
+    Validate model
+
+    Raises
+    ~~~~~~
+    ValueError is model is None
+
+    # TODO: Should we add ModuleNotSupportedError?
+    """
+    if model is None:
+        raise ValueError("Model is none")
